Complete transactions even if we're not actively waiting for them.
Instead of only marking the transactions we're currently waiting for
as complete, mark any transactions that may complete as done when they
finish in the busy loop.

This fixes a potential hang when endpoint A does a write, followed by
endpoint B, followed by endpoint A. If we get the completion from
hardware for B before A, then the final wait in A will never complete.