Allow setting the ‘wMaxPacketSize’ on an endpoint.
Adds a new macro ‘EPTYPEANDMAXLEN’ for setting the max packet size on
a given endpoint by encoding it in the bottom 8 bits of the
corresponding ‘epBuffer’. The existing macro for ‘EPTYPE’ defaults to
using ‘USB_EP_SIZE’ so no code needs to change if that's adequate.

Use the new macro in the CDCACM device on the ACM endpoint, which
specifies a 16 byte packet length.