Some code re-factoring by using ‘constexpr’ instead of macros.
This gets us some ability to debug a little more easily on gdb, since
gdb can't deal with #defines. This also gets us more static type
checking, although there are still some casts so it's not perfect.

Also move the store for ‘epBuffer(unsigned int)’ to the ‘EPBuffers()’
singleton, which is specialized on the number of endpoints available.