FlashStorage: Fix the bounds checking in FlashStorage::write()
The bounds checking in `FlashStorage::write()` was insufficient, it only checked
whether the offset is out of bounds, or if the data size is bigger than the
storage space. It did not account for the case where offset is within bounds,
data size is smaller than storage size, but there isn't sufficient space for
data size starting at offset.

To fix that, correct and simplify the bounds checking: we'll bail out if `offset
+ data_size` is larger than `_storage_size`. This covers both previous cases
that were covered, and the third case too, which previously was not.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>