fix receiving ZLP
Correctly handle zero-length packets by re-enabling receive. The API
doesn't have a reliable way to signal receiving data of zero length.
This meant that the application might not ever read the (empty) data,
causing the endpoint to lock up and stop accepting more data.
Re-enabling receive "early" for ZLPs fixes this problem.

Some hosts send a zero-length packet when sending data that's a multiple
of wMaxPacketSize (64 bytes on full-speed USB). This includes at least
some versions of macOS. USB generally uses short packets as a signal of
end of transmission, which includes sending ZLPs if the data is a both
multiple of wMaxPacketSize and less data than the receiver was
expecting. Because the host might not know the size of the device's
receive buffer, it might send a ZLP following a transmission that ends
with a max-length packet, to ensure that the device receives it
promptly.

Signed-off-by: Taylor Yu <code@argon.blue>